requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
Django==4.2.7
django-redis==6.0.0
djangorestframework==3.14.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
redis==6.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
